refs #294 Fix migrations

+++ server/trunk/db/migrate/20110424185724_create_locations.rb
@@ -1,8 +1,8 @@
 class CreateLocations < ActiveRecord::Migration
   def self.up
     create_table :locations do |t|
-      t.double :latitude
-      t.double :logitude
+      t.float :latitude
+      t.float :logitude
 
       t.timestamps
     end
